refactor(slack-bot): extract claude-cli.js from index.js (behavior-preserving)
